+++ src/Routers/AggregateRouter.js
@@ -91,21 +91,20 @@ export class AggregateRouter extends ClassesRouter {
 
   static transformStage(stageName, stage) {
     if (stageName === 'group') {
-      if (Object.prototype.hasOwnProperty.call(stage[stageName], '_id')) {
+      if (Object.prototype.hasOwnProperty.call(stage[stageName], 'objectId')) {
+        // TODO: show deprecation warning for `objectId`
+        stage[stageName]._id = stage[stageName].objectId;
+        delete stage[stageName].objectId;
+      }
+      if (!Object.prototype.hasOwnProperty.call(stage[stageName], '_id')) {
         throw new Parse.Error(
           Parse.Error.INVALID_QUERY,
-          `Invalid parameter for query: group. Please use objectId instead of _id`
+          `Invalid parameter for query: group. Missing key _id`
         );
       }
-      if (!Object.prototype.hasOwnProperty.call(stage[stageName], 'objectId')) {
-        throw new Parse.Error(
-          Parse.Error.INVALID_QUERY,
-          `Invalid parameter for query: group. objectId is required`
-        );
-      }
-      stage[stageName]._id = stage[stageName].objectId;
-      delete stage[stageName].objectId;
     }
+
+    // TODO: show deprecation warning for missing `$`
     const key = stageName[0] === '$' ? stageName : `$${stageName}`;
     return { [key]: stage[stageName] };
   }